Fix hang when requesting report from fully compliant enclave.
When the View->cOnnections menu item was selected for a remote
enclave that Intel was reporting an OK attestation status the
cboot-mgr utility would hang.  This was secondary to the fact
that the CbootManager->get_connection() method was attempting to
match the following line:

^.*Extended group id.*$

Which is only displayed when a GROUP_OUT_OF_DATE attestation
response is returned from IAS.  When an OK attestation response
is provided this string is not displayed and the client thus
hangs.

This was corrected by attempting to match the following line:

^Platform status: OK$

In addition to the 'Extended group id' string.  This results in
the close of the query for information on the status of the
connection.