Add directory for documentation.
As the initial commit in a version that will focus on a transition
to new libraries and a general code cleanup the Documentation
directory was added to support documentation for the NAAAIM
system.

The 'Changes' file in this directory will be used to log one line
documentation on major changes implemented in the new version.

Starting with this release a transition was made to a new
hardware based GPG signing key.  The Keys file documents the
transition from the old signing key to the new signing key that
will be used in subsequent commits.