Implement enclave exception handler.
This commit implements a major update to the runtime
functionality in the form of formal enclave exception handling.

Enclave exceptions are unique in that they may be either a fatal
exception or a request for services to the untrusted runtime.
The best example of the latter is a request to handle an
instruction, such as CPUID, that is prohibited while the process
is running in enclave mode.

A processor exception while running in enclave mode is treated as
an Asynchronous Enclave eXit (AEX) condition.  The exception
handler address is specified to the enclave when the initial
EENTER instruction is executed for an enclave.

If an AEX condition is generated the processor registers are set
to a synthetic state that is defined in the Intel Software Development
Manual.  Execution begins at the exception handler address with
generation of the appropriate signal for the untrusted runtime.

The signal handler determines whether or not it will choose to
handle the exception.  If so the signal context is examined in
order to locate the Task Control Segment pointer, the OCALL
dispatch pointer and the address of the object describing the
enclave.

These pointers are used to setup a return call to the enclave
with a request to execute the exception handler, if any, that the
enclave has defined.  The exception handler is responsible for
analyzing the exception and determining whether or not normal
execution should be resumed.  If a decision is made to continue
execution an SGX_SUCCESS value is returned to the second
invocation of the enclave.

Upon receipt of an indication that the exception handler has been
successfully executed the signal handler will return which will
re-start execution inside of the enclave.  If not the signal
handler will abort execution of the program.

An additional function was added to the boot-sgx.s which contains
the enclave entry code.  The function,
boot_set_get_exit_handler(), returns the address of the enclave
exception handler.  The exception handler uses this to determine
if this is an enclave based exception.  It does so by comparing
the address in the RIP register with the address of the exception
handler.

In order to configure the exception handler an enclave handling
application must call the sgx_configure_exception() function.
This function will return a boolean false value if the exception
handler cannot be configured.