Update Intel SDK and OpenSSL releases.
This update implements changes needed to compile against the 2.4
release of the Intel SDK.  In addition this release compiles
agains the 1.1.0j release of the SGX version of OpenSSL.

The atexit() function has been integrated into the Intel supplied
C library.  The prototype and null definition for the function
was deleted from the SGXfusion shim code.  It is unclear whether
or not the atexit() implementation will properly detect improper
pairing of SGXfusion constructor and destructors.

The unit test for the SGXnaaaim library is modified to register
an SGX exception handler so that the OpenSSL library will
properly initialize its platform specific CPU behavior.  The
exception handler needs to be propagated to other enclaves that
are using OpenSSL functionality.