Automatically setup enclave exception handling.
The SGXenclave->setup() method implements a wrapper function that
implements all of the functionality needed to bring an enclave to
executable status.  Given this intent it is logical to have this
function also include the call to initialize the enclave
exception handling infrastructure.

If the developer chooses to implement the independent setup of an
enclave the expectation is that they will need to include a call
to the function that sets up the exception handler.

Since multiple enclaves may be loaded by an application an
exclusion flag is implemented in the exception handler that
prevents the handler from being executed more then once in the
context of a single process.