Drop user space alignment of enclave image.
At one time their was an expectation by the SGX device driver to
have the platform software align the enclave image.  This is no
longer the case and continueing to do so wastes EPC memory.

This update drops the alignment of the image and leaves it to the
discription of the driver.