Add exception handler to Duct unit test driver.
Secondary to the fact that all of the applications above the
SGXnaaaim library level link against the SGXSSL library they need
to have support for exception handling.  The test-Duct driver
uses the staged enclave setup strategy and thus will not
implicitly setup the exception handler.  This commits add the
necessary setup call.