Add non-token mode to enclave loader unit test.
The proposed mainline kernel driver does not allow a token to be
passed to the EINIT ioctl.  For testing purposes the -N switch
was added to the sgx-load utility in order to request non-token
initialization.

Additional verbosity was added to each step of the enclave
initialization procedure in order to provide diagnostics on when
and how the setup of an enclave may have failed.