Adjust size calculation for non-zero segments.
An update to the Intel 2.4 signed binaries demonstrated that the
heuristic for computing the number of null pages to be added for
a segment with a non-zero offset was still incorrect.  The
platform certification enclave, libsgx_pce.signed.so, would not
properly initialize.  The remainder of the signed enclaves would
load with the current heuristics.

After some experimentation a modified heuristic was found which
loads in proper initialization of all the Intel supplied binary
enclaves and passes the RDK regression test.  This change
involves checking whether or not the size of the loaded segment
is larger then the memory size plus the segment offset.  If so
the size variable is set to the difference between the size of
the currently loaded segment and the memory_size+offset value.

The condition which decides whether or not a page should be
subtracted from the null page size was modified to carry out the
reduction if the memory size of the segment is less then the
loaded size.

I fear that this may not be the last change in the heuristic that
will be needed.