Add support for token-less enclave execution.
In order to support what appears to be the design of the SGX
driver intended for inclusion into the Linux mainline sources
this commit adds support for initialization of an enclave without
an EINITTOKEN.

At the time of this writing the new Linux driver does not allow a
pointer to an EINITTOKEN to be specified in structure that is
passed to the EINIT ioctl call.  We have a very straight forward
patch for the driver that allows a pointer to be specified but
returns EINVAL if a non-null pointer is provided.  This allows
the RDK to function with either the out-of-tree or the in-tree
kernel driver.

The SGXenclave object was modified to test whether or not a NULL
EINITTOKEN value was specified and if so does no update the SECS
object that is maintained in the object be updated with the
MRSIGNER, MRENCLAVE and CPUSVN values.  This will cause quote
generation to fail since generation of a quote requires obtaining
the CPUSVN value from the initialization of the PCE enclave.

It thus appears that remote attestation will not work with the
proposed in-kernel driver.

The SGXenclave->setup() method was also modified to handle either
a NULL pointer being passed as the name of the file containing an
initialization token.  Also acceptable is a null string
consisting of a single null character.  This allows a ""
arguement to be passed to any utility which uses the -t switch to
specify an initialization token.

The sgx-load utility was modified to follow a similar strategy
since it uses incremental enclave preparation rather then the
consolidated SGXenclave->setup() method.

The test-naaaim and test-Duct utility were modified to use a
similar strategy since they use incremental setup.  The
test-Possum utility uses the consolidate setup and should need no
modification.