Add support for creating statically linked unit test drivers.
In order to efficiently generate the RDK test kit the ${STATIC}
variable was added to the LDFLAGS variable definition in the
Makfiles for the SGXfusion and SGXnaaaim libraries.  This allows
the STATIC=-static variable definition to be given at the top of
the SGX heirarchy to product statically linked unit test drivers.