Implement single location for enclave signing key.
Due to cutting and pasting of enclave build directives each
directory that was building an enclave had ended up with its own
private signing key.  This made it difficult to switch to signing
with a production key.

There is also a great deal of replication in the Makefiles that
are building enclaves of commonly defined variables such as the
definition of the enclave link command.

The strategy that was initiated was to place all enclave relevant
definitions and declarations that are commonly used into a file
name Make.sgx in the top level of the SGX build heirarchy.  Each
Makefile then uses a relative reference to this file to pick up
its definitions.

The debug signning key was moved to a file named as follows:

Signing-key.pem

Which also resides in the top level SGX build directory.  The SGX
specific make include file optionally sets the SGXKEY variable.
This allows a top level make in the SGX directory to use the
following invocation:

make SGXKEY=

To set an alternate definition for the location of the signing
key.

The definition for the location of the signing tool was also
consolidated in the include file as was the location of the Intel
SDK directory.

There will be further work to consolidate these definitions in
subsequent commits.