Implement support for non-debug enclaves in unit tests.
This commit modifies the test-fusion, test-naaaim and test-Duct
unit tests to allow the selection of whether or not the enclave
is to be initialized and run in non-debug, ie. production mode.

The -d switch which had previously set the debug state of the
enclave was modified to set PSW debug mode for the loading and
initialization of the enclave.  This switch had previously
selected whether or not the enclave was to be initialized in
debug or production mode.

The -p switch, as is the case with the sgx-gen-token utility,
requests that the debug attribute be removed from the enclave.
This causes the enclave to be run in production mode.

The token generated for the enclave has to match the mode the
enclave is to run in.  If the unit test is run with the -p switch
the enclave token has to have been generated with the -p switch
to the sgx-gen-token utility.