Migrate SHA256 object to NAAAIM library.
Secondary to the history of this project the object implementing
SHA256 hashing was in the top directory of the source heirarchy.
This commit moves the object into the NAAAIM library directory
and integrates it into the archive object.

This also requires changes throughout the build structure to
remove direct references to the SHA256.o object file.