Use ENCLAVE_DIR to set enclave location for cboot and cboot-mgr.
The changes to the unit tests which implemented the use of the
ENCLAVE_DIR CPP to set the directory location of the unit test
enclaves ended up breaking the compilation of the cboot and
cboot-mgr utilities.

The modifications to the build heirarchy are not yet complete nor
robust with respect to the handling of this issue.  Until that
work can be completed a workaround was implemented which
explicitly checks for the ENCLAVE_DIR and if it is not set a
value consistent with the current behavior of the utilities is
set.

This allows the ENCLAVE_NAME CPP directive to be used without
compilation errors in these utilities.