Convert SHA256 object to new object name.
The 1.1.x release of OpenSSL implements an API change that
conflicts with current SHA256 naming convention since a function
named SHA256() was implemented in this release to simplify the
computation of hashes.

In order to prepare for this the SHA256 object was renamed to
Sha256.  This commit makes the name change in the object itself
and all of the instances that were referenced.