Update SSLDuct to OpenSSL 1.1.x API.
This commit updates the object that implements SSL/TLS based
communications so that it compiles under the OpenSSL 1.1.x API.
A change to the standardized copyright declaration was also made.

The previous object used the SSLv3 protocol which has long been
deprecated.  This update converts the object to select the
default TLS implementation for both the client and server ends
of a context.

The object contains a custom generated Diffie-Hellman key.  The
OpenSSL API change no longer allows the P and G parameter fields
of the key structure to be directly manipulated.  The code now
uses the DH_set_pqg() function to set the key parameters.

It is important to note that these are only compilation fixes,
the object needs to undergo testing to verify the correctness of
these changes.