Convert RSA key handling to OpenSSL 1.1.x API.
The API changes in the 1.1.x release of OpenSSL focused on making
objects opaque.  Accessor functions were added to allow values in
these opaque structure to be set and retrieved.

The object which manages RSA keys for the SGX RDK code was
modified to replace direct access to the key elements with the
appropriate accessor function.

The EPID provisioning utility was also modified to use the access
functions to set the RSA key elements.