Fix linkage of utility binaries.
In order to avoid special casing of the linker commands needed to
create applications linked against the MUSL C library a
modification to the custom specifications file was modified in
Wok version 0.4 to explicitly specify the RPATH value that
applications are to be created with.

This allowed the RPATH specification to be dropped from the
LDFLAGS variable.  This will be extended to other Makefiles as
well.

A specific issue was noted with the tpm2-cmd which links against
the libtss library.  That library had specific dependencies
against the libcrypto.so.1.0.0 shared library which caused
linkage errors when the tpm2-cmd binary was linked against the
libcrypto.so.1.1 shared library.

To correct this the tpm2-cmd was explicitly linked against the
libcrypto.so.1.0.0 shared library.

There are also issues with applications that are linking against
the TPM1.2 libtspi library.  Due to the limited use of these
binaries this condition will be allowed to stand.