Fix uninitialized variable in PossumPipe.
Due to code being commented out an unitialized variable warning
was being issued.  If the commented out code path was active
there would be no reason to not have the variable initialized.
In order to attain the current goal of no compilation errors the
variable was initialized at its declaration.