Eliminate MUSL link arguement dependencies.
Additional time was spent investigating how the MUSL C library
searches for shared libraries when an application is loaded.  The
GNU C library used the /etc/ld.so.conf file to specify a set of
directories which the ldconfig utility uses to build the
/etc/ld.so.cache file that the dynamic linker uses to resolve
library dependencies.

The MUSL C library used the /etc/ld-musl-x86_64 file to specify a
set of directories which are to be searched at load time to
resolve library dependencies.  This file expects a set of colon
separated directory locations.

The Wok development environment was modified to include the
configuration file with a single entry specifying the
/usr/local/musl/lib directory.  With this modification in place
there is no need to specify the '-Wl,-rpath /usr/local/musl/lib'
arguement to the linker in order to code a runtime path into the
binary.

This lays the framework for additional build modifications to
allow a build with GCC with minimal configuration changes.