Generate launch enclave whitelist from SGX_runtime.
The launch enclave whitelist was being generated from a copy that
had been downloaded from the following URL:

https://whitelist.trustedservices.intel.com/SGX/LCWL/Linux/sgx_while_list_cert.bin

Since this was not integrated into the source code repository the
build would fail on a virgin installation of the source.

A copy of the binary whitelist was integrated into the
SGX_runtime environment at the following path:

/opt/intel/sgxpsw/aesm/sgx_white_list_cert.bin

The include file that defines the white list to the runtime is
built from this binary file.