Automatically embed launch enclave.
The standalone binary implementations were referencing an
ASCII based implementation of an extremely dated Intel binary
launch enclave.  Following the model of the strategy used to
embed the whitelist the build system was modified to
automatically reference the currently installed binary launch
inclusion to generate an ASCII image of the enclave.

This ASCII image is named LE.h and the two standalone binaries
were modified to include this file in order to create an array
that can referenced by the SGXenclave->open_enclave_memory
method to load the enclave.

At the current time neither binary is functional due the code not
having been maintained for some time.  These fixes are
implemented in order to have a fully independent and functional
build of the RDK.

When this code is freshened the proper strategy will be to follow
the model used for other enclaves and wrap the launch enclave in
its own object which will make standalone binaries require much
less replication of code.