Automatically embed test enclave.
The final step in cleaning up the build of standalone binaries
was to automatically embed the test enclave in the unified test
binary.

Since the test-ecall.signed.so binary had been committed to the
source tree that signed bary was converted to ASCII form in the
build process so that it could be included as the TE.h include
file.

Once again these two utilities are not functional due to some
type of loader error in the launch enclave.  The
test-ecall.signed.so binary needs to be dropped from the tree and
replaced with an enclave that is generated as a component of the
build process.