Only configure and build HurdLib when required.
The build change to automatically build the HurdLib support
library as a component of a normal build was needlessly running
the configure and build process even when the library had been
built.

This update modifies the build process to detect if the library
has been built and does not invoke the build process if this is
the case.