Handle platforms that have been previously provisioned.
The provision-epid script did not properly handle platforms that
have been previously provisioned.  This commit provides support
for using this script to re-provision platforms that have had
initial EPID provisioning conduct on them.

The SGXmessage->get_response_type method was implemented in
order to determine the message type that is returned from the
Intel provisioning service.  This function was used to implement
a 'test' mode for the sgx-provision utility that can be used to
determine the type of a previously captured message.

The provision-epid script was significantly upgraded to use the
test mode to determine if the platform has been previously
provisioned and modifies its execution path accordingly.

Option processing was also added to interpret the -o option as a
request to change the location where the EPID is placed.  By
default the output location is:

/var/lib/IDfusion/data/EPID.bin

To match the expectations of the runtime.