Implement silent provisioning by default.
The sgx-provision utility was generating verbose output even when
the -v arguement was not given on the command-line.  The flag
variable was moved into static context for the file so that it
could be referenced by all functions participating in the
provisioning process.  Additional conditionals were added to
fence off output generating portions of the code unless the
Verbose variable was set to a true value.

The provision-epid script was modified to take a -v command-line
arguement that triggers the specification of the -v command-line
arguement on all invocations of the sgx-provision binary.