Set defaults for remote attestation unit test.
The test-remote-target utility that provides unit testing for
both untrusted and trusted implementations of remote attestation
quoting required explicit setting of the location of the sealed
EPID as well as the initialization tokens for the QE and PCE
enclaves.  This made use of the utility somewhat cumbersome and
potentially error prone for users inexperienced with the use of
the RDK.

This update sets standard default locations for these entities
which allows the utility to be run in test mode without
specification of any arguements.