Allow specification of signature structure.
The generation of the ASCII representation of the signature
structure for the Intel launch enclave was being taken from the
following fixed file location:

/opt/intel/sgxpsw/aesm/le_prod_css.bin

Work on testing an independently signed version of the launch
enclave indicated that it would be necessary to reference an
alternate signature structure.  The following variable was
created to define the location for the signature structure file:

LE_SIGSTRUCT

When was then used as an arguement to the generate-array utility
that converts the binary file into a character array for
inclusion in the binary.  This allows a command-line override of
the variable to specify an alternate signature structure.