Update to version 3 of Intel Attestation Services API.
In the fall of 2018 in response to the L1TF vulnerability Intel
deprecated version 2 of their Intel Attestation Services (IAS)
API in favor of version 3.  The end of life date for the version
2 API is 4/2/2019.

The primary changes in the version 3 API was to implement
infrastructure for reporting on the status of hyperthreading support
on the platform, most specifically its enablement in the BIOS.
If hyperthreading status changes are needed to ensure platform
security IAS returns CONFIGURATION_NEEDED as the reason for a
GROUP_OUT_OF_DATE response.

This update implements the minimal set of changes required to
convert to the version 3 API.  Most notable is the change in the
REST service URL and the addition of an enumeration type and
description string for the CONFIGURATION_NEEDED status.