Implement IAS version verification.
This commit modifies the SGXquote->decode_report method so that
it verifies the version of IAS that is generating an attestation
report.  The object verifies that the report is consistent with
the IAS version that the object was designed to interact with.

The publication of a report version is a new feature introduced with
the Intel IAS version 3 API.  Intel recommends that a report
verifier confirm the presence of a report version and its value as a
component of verifying the status of an attesting platform.