Port enclave based SGXquote to IAS version 3 API.
This update adds the changes needed to implement the processing
of IAS version 3 API attestation reports to the enclase based
SGXquote object.

In addition to the API changes the _init_state function was
updated to properly initialize the object state variable.  The
object destructor was not properly destroying all of the objects
in the SGXquote state object and the ->whack method was updated
to do so.