Surface secure endpoint connection parameters.
This update adds the PossumPipe->get_connection method which
provides an accessor method for retrieving status information on
the security parameters of the remote endpoint.  It is provided
as an interim step to a more sophisticated method for determining
whether or not communications should be enabled with the remote
endpoint.

The following information is currently returned from the
PossumPipe object:

	Enclave signer
	Enclave measurement
	Enclave attributes
	Enclave vendor identifier
	Enclave vendor security version

The PossumPipe unit test enclave was updated to print out this
information on the server and client sides after a successful
connection attempt.