Fix return code and format for sgx-gen-verifier utility.
The utility was not properly setting the return code to 0 when an
identity verifier file was being written out.  This update
sets the return value correctly if there are no errors
encountered in writing out the identifier verifier.

The format for the identifier filename was changed so that an
underscore is used to separate the default identity verifier which
is of the form:

SHA256.ivy

From the label value provided as an arguement to this utility.

This is designed to implement the following tagged format that
will be used by the provision-verifiers utility that will be
subsequently committed:

SHA256.ivy_HOSTNAME.ENCLAVENAME