Implement identity verifier provisioning utility.
The update provides an implementation of the following utility;

provision-verifier

That is used to simplify the creation and installation of
identity verifiers to support IDsecure/POSSUM communication
pipes that implement Host Specific Enclave Authentication (HSAE).

This utility assumes that a utility that uses a specific enclave
follows a standard format where the -M arguement is used to
request that the utility execute an ECALL to have the enclave
generate an an SGX MRSIGNER derived key and output its value.

This value is embedded in the identity verifier so that a client
will be able to confirm that the secured endpoint is one that has
generated the verifier being used.

The utility operatins either in generate or install mode
specified by the -G and -I command-line switches respectively.
The generate mode causes an identity verifier to be written to a
file in the following form:

SHA256.ivy_HOSTNAME_ENCLAVENAME

By convention any infrastructure using a PossumPipe object loads
all identity verifiers of the following form:

/opt/IDfusion/etc/verifiers/ENCLAVENAME/*.ivy

The install mode of the utility processes a filename in the form
generated by this utility to determine where to install the file.
The utility then creates a symbolic link that points from the
locations that the PossumPipe utilities will use to the actual
filenames.