Implement quiet mode for identity verifier generation.
The -q command-line switch was implemented for the
sgx-gen-verifier utility to suppress output of the contents of
the identity verifier.

The provision-verifier utility was then modified to use this
switch on its invocation of the the sgx-gen-verifier utility in
generate mode.  This significantly reduces the amount of output
in what will be the most common method of generating identify
verifiers.