Implement provision-tokens utility.
This update adds the provision-tokens utility.  This utility provides
convenience functionality primarily for installation and upgrades
of the SRDE runtime.

The utility generates launch tokens for all of the enclaves
located in the following directory:

/opt/IDfusion/lib/enclaves

The tokens are placed in the following directory:

/var/lib/IDfusion/tokens

In the following form:

ENCLAVENAME.token

Which is the expected format for any applications using SRDE
standard pathname specifications.