Update cboot exception handler.
The cboot utility was failing due to an illegal instruction
fault.  This is secondary to the fact that the cboot utility was
installing its own signal handler and was using manual setup of
the enclave rather then using the SGXenclave->setup method.

Since the cboot utility needs to have its own signal handlers
installed the cboot signal handler was modified to use SIGINFO
format.  The strategy was to have the cboot signal handler test
and handle the signals it was tasked with handling before calling
the SGX exception handler.  This provided the SGX exception
handler an opportunity to service its signals which included
calling back into the enclave to handle the Asynchronous Enclave
Exit events.

The setup of the cboot signal handler was modified to include
registration of the signals that the SGX handler needs to handle
in addition to requesting SIGINFO style handling.