Handle socket close in Duct object.
A PossumPipe object in server mode was not elegantly handling a
situation where the client end of the connection was interrupted
with a SIGINT signal (CNTRL-C).  This caused the error abort
pathway to be taken on the server side of the PossumPipe
connection.  This typically caused an exit of the application.

At the Duct protocol level a send of 0 bytes was being
interpreted as an end of session directive and the eof boolean
value in the state variable was being set.  This handled an
application signaling an end of a network session but unless a
signal handler was installed to process SIGINT this method of
signalling the end of the session was inadequate.

An abrupt socket close by a client is detected by a zero-length
read of the length header for a socket.  The code was currently
selecting an error exit when a return value of anything other
then four bytes was encountered.  The POSIX definition for the
read indicates that a return value of 0 signals an end of file
condition.

The action portion of the conditional was modified to check if
the return count from the read was 0.  If so the Duct instance
is tagged as having an EOF condition and a non-exiting error
condition set.

The EOF condition was being properly propagated upward into
the enclave instance of a Duct object.  The PossumPipe utility
was modified to check for the EOF condition on the Duct object
providing I/O support for the PossumPipe.  If the Duct has
an EOF condition a non-error exit was taken with an end-of-pipe
(EOP) status.