Implement PossumPipe internal error tracking.
This update improves error handling in the fact of invalid
identity verifiers.  Previously any error in the setup of a
PossumPipe conduit would result in an error exit from the
application.  This was particularly troublesome for server
applications which were designed to stay operational in the face
of various types of client behavior.

A new enumerated data type was created to manage the various
errors states being implemented.  The following four errors are
initially defined:

PossumPipe_error_internal

PossumPipe_error_closed_pipe

PossumPipe_error_no_identity

PossumPipe_error_invalid_identity

By default the PossumPipe error state is set to internal error.
This is a catch all error type for any type of error with a low
probability of occurrence.

The closed pipe directive is used to indicate that the
communiations counter-party has chosen to close the socket
connection.  This is used to detect a situation where the
communiations counter-party is unwilling to accept further
communications from the client.

The 'no identity' error is used to indicate that a matching
identity verifier was not found that matched the host
initialization challenge.

The 'invalid identity' error indicates that a matching identity
verified was found but it was unable to decrypt the host
initialization challenge.

The PossumPipe->get_error method as added to obtain the value of
the internal error value.

The PossumPipe object was updated to set these internal states
depending on the conditions that was detected.  If these states
are detected an error return is taken but without printing out an
error message for that condition.  This allows the caller to more
elegantly print out the error state.

Adding this error handling exposed a problem with the object
destructor.  The instructor assumed that all object had been
initialized which is not the case when these early exits are
taken.  Null object tests were implemented for internal state
objects that are not guaranteed to have an active implementation.