Update SGX capability tester.
The sgx-check utility that outputs the platform SGX capabilities
was updated to print out the values of the launch control
registers.  These are four 64-bit registers that contain the
SHA256 hash of the value of the key that is allowed to authorize
the initialization of enclaves.

By default this is set to the signature of the key that Intel
uses to sign the launch enclave.  On a Flexible Launch Control
platform these registers can be written to allow multiple keys
to authorize the initialization of enclaves.

The output of the miscellaneous select bits was not being
properly indented and this was corrected as well.

A more informative program banner was also implemented.