Support compilation under GCC.
The SRDE was originally designed to build under musl-gcc.  As a
result it had dependencies that prevented the source base from
being compiled under a standard GCC implementation.

This commit makes the necessary and minimal changes needed to
support compilation under GCC.

Handling dependencies for the SRDE is somewhat problematic since
the dependencies include both trusted and untrusted
implementations.  This is necessitated by the fact that the the
NAAAIM and HurdLib libraries are designed to compile from common
code in both the trusted and untrusted variants.

One of the problematic issues to date has been Support
compilation under GCC.

The SRDE was originally designed to build under musl-gcc.  As a
result it had dependencies that prevented the source base from
being compiled under a standard GCC implementation.

This commit makes the necessary and minimal changes needed to
support compilation under GCC.

Handling dependencies for the SRDE is somewhat problematic since
the dependencies include both trusted and untrusted
implementations.  This is necessitated by the fact that the
HurdLib and NAAAIM libraries are designed to compile in unchanged
form for both standard userspace and enclave implementations.

An example of one of the problematic dependencies is OpenSSL.
The SGXSSL implementation is based on OpenSSL 1.1 while OpenSSL
1.0 is very common in standard Linux distributions.  Since there
are API differences between 1.1 and 1.0 there is a need so
support OpenSSL 1.1 irregardless of what implementation version
the Linux distribution provides.

The strategy for supporting dependencies is to use a common
library implementation root.  This is inspired in no small part
by the fact that the MUSL implementation currently roots all of
its binaries and libraries in the /usr/local/musl heirarchy.

For the purposes of the GCC implementation the
/usr/local/IDfusion root was selected.  This allows alternate or
non-distribution supplied libraries to be rooted in a heirarchy
that does not conflict with existing libraries.

Additional changes that were needed include adding include file
definitions that GCC requires that the MUSL include
infrastructure seems to automatically pick up.  The primary
issue was the need to specifically include stdint.h in the GCC
based compiles.

One of the other differences was is that the MUSL regex library
defines REG_OK as a result type while glibc uses REG_NOERROR.
This was addressed by defining REG_OK as REG_NOERROR if REG_OK
was not defined.

Due to the somewhat dated libc implementation that is supporting
the GCC implementation there was a need to include a definition
for the MS_MOVE constant.  Once again this was supplied by a file
specific define if the MS_MOVE definition was not supplied by
default.

One of the issues that needs to be addressed are definitions for
the location of the OpenSSL libraries.  Due to expediency needs
the strategy was to define the OpenSSL library and include
definitions by checking the version of CC that is being used.  If
musl-gcc is the C compiler the /usr/local/musl heirarchy is used.

All of this underscores the need for improvements to the current
build infrastructure.