Implement corrected C++ guard macro spelling.
Testing of an initial C++ based SRDE application indicated that
the linkage was failing secondary to the ability to resolve
references to various HurdLib based objects.  This was traced
back to a mispelling of the __cplusplus guard macro that was
gating definition of the HCLINK macro to define C linkage for the
library objects.

This update pulls in an updated version of HurdLib that defines
the macros properly.  The version of the HurdLib.h header file
for the creation of applications in enclave context was updated
as well.