Implement C++ unit test driver for the SRDE fusion library.
This commit implements a C++ based version of the driver utility
that tests the functionality of the fusion library that provides
the compatibility layer for enclaves that use the HurdLib object
library.

This unit test verifies the compatibility work that has been done
to allow C++ based applications to use the SRDE framework.