Add TEEnaaaim unit test driver.
This update introduces a C++ based version of the unit test
driver that exercises the functionality of the trusted execution
environment version of the NAAAIM library.

This required updates to code outside of the trusted version of
the NAAAIM library.  Some of this changes were where C linkage
definitions were required that had not been detected in early
commits which introduced this functionality.